feat: Darkmode can now be toggled everywhere
Darkmode can now to toggled outside the darkmodeSwitcher component. It does this using a store that gets read and written to by the darkmodeSwitcher
