+++ src/lib/stores/theme.ts
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const isDarkmode = writable(null);
